feat: Sent Module Response DTO SSNetwork으로 변경
