+++ Projects/Core/SSNetwork/Sources/DataModel/CategoryModel.swift
@@ -1,8 +1,8 @@
 //
 //  CategoryModel.swift
-//  Received
+//  SSNetwork
 //
-//  Created by MaraMincho on 7/4/24.
+//  Created by MaraMincho on 8/22/24.
 //  Copyright © 2024 com.oksusu. All rights reserved.
 //
 
@@ -10,17 +10,17 @@ import Foundation
 
 // MARK: - CategoryModel
 
-struct CategoryModel: Codable, Equatable {
+public struct CategoryModel: Codable, Equatable {
   /// 카테고리 아이디
-  let id: Int
+  public let id: Int
   /// 카테고리 순서
-  let seq: Int
+  public let seq: Int
   /// 카테고리 이름
-  let name: String
+  public let name: String
   /// 카테고리 스타일
-  let style: String
+  public let style: String
   /// 기타
-  let isMiscCategory: Bool
+  public let isMiscCategory: Bool
 
   enum CodingKeys: String, CodingKey {
     case id
