--- Projects/Feature/Received/Sources/Commons/FriendModel.swift
@@ -1,10 +1,0 @@
-//
-//  FriendModel.swift
-//  Received
-//
-//  Created by MaraMincho on 7/4/24.
-//  Copyright © 2024 com.oksusu. All rights reserved.
-//
-
-import Foundation
-
