--- Projects/Feature/Received/Sources/Commons/GetFriendStatisticsResponse.swift
@@ -1,10 +1,0 @@
-//
-//  GetFriendStatisticsResponse.swift
-//  Received
-//
-//  Created by MaraMincho on 7/5/24.
-//  Copyright © 2024 com.oksusu. All rights reserved.
-//
-
-import Foundation
-
