--- Projects/Feature/Received/Sources/Commons/PageResponseDtoGetFriendStatisticsResponse.swift
@@ -1,10 +1,0 @@
-//
-//  PageResponseDtoGetFriendStatisticsResponse.swift
-//  Received
-//
-//  Created by MaraMincho on 7/5/24.
-//  Copyright © 2024 com.oksusu. All rights reserved.
-//
-
-import Foundation
-
