+++ Projects/Feature/Received/Sources/LedgerDetail/LedgerDetailMain/LedgerDetailProperty.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import SSNetwork
 
 // MARK: - LedgerDetailProperty
